Regroup native classes of crypto module
