+++ stdlib/crypto/src/main/java/org/ballerinalang/stdlib/crypto/nativeimpl/Decrypt.java
@@ -28,11 +28,42 @@ import java.security.PrivateKey;
 import java.security.PublicKey;
 
 /**
- * Extern function ballerina.crypto:decryptRsaEcb.
+ * Extern functions ballerina decrypt algorithms.
  *
  * @since 0.990.4
  */
-public class DecryptRsaEcb {
+public class Decrypt {
+
+    public static Object decryptAesCbc(ArrayValue inputValue, ArrayValue keyValue,
+                                       ArrayValue ivValue, Object padding) {
+        byte[] input = inputValue.getBytes();
+        byte[] key = keyValue.getBytes();
+        byte[] iv = null;
+        if (ivValue != null) {
+            iv = ivValue.getBytes();
+        }
+        return CryptoUtils.aesEncryptDecrypt(CryptoUtils.CipherMode.DECRYPT, Constants.CBC, padding.toString(), key,
+                input, iv, -1);
+    }
+
+    public static Object decryptAesEcb(ArrayValue inputValue, ArrayValue keyValue, Object padding) {
+        byte[] input = inputValue.getBytes();
+        byte[] key = keyValue.getBytes();
+        return CryptoUtils.aesEncryptDecrypt(CryptoUtils.CipherMode.DECRYPT, Constants.ECB, padding.toString(), key,
+                input, null, -1);
+    }
+
+    public static Object decryptAesGcm(ArrayValue inputValue, ArrayValue keyValue,
+                                       ArrayValue ivValue,  Object padding, long tagSize) {
+        byte[] input = inputValue.getBytes();
+        byte[] key = keyValue.getBytes();
+        byte[] iv = null;
+        if (ivValue != null) {
+            iv = ivValue.getBytes();
+        }
+        return CryptoUtils.aesEncryptDecrypt(CryptoUtils.CipherMode.DECRYPT, Constants.GCM, padding.toString(), key,
+                input, iv, tagSize);
+    }
 
     public static Object decryptRsaEcb(ArrayValue inputValue, Object keys, Object padding) {
         byte[] input = inputValue.getBytes();
@@ -46,6 +77,6 @@ public class DecryptRsaEcb {
             return CryptoUtils.createError("Uninitialized private/public key");
         }
         return CryptoUtils.rsaEncryptDecrypt(CryptoUtils.CipherMode.DECRYPT, Constants.ECB, padding.toString(), key,
-                                             input, null, -1);
+                input, null, -1);
     }
 }
