--- stdlib/crypto/src/main/java/org/ballerinalang/stdlib/crypto/nativeimpl/EncryptAesEcb.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.stdlib.crypto.nativeimpl;
-
-import org.ballerinalang.jvm.values.ArrayValue;
-import org.ballerinalang.stdlib.crypto.Constants;
-import org.ballerinalang.stdlib.crypto.CryptoUtils;
-
-/**
- * Extern function ballerina.crypto:encryptAesEcb.
- *
- * @since 0.990.4
- */
-public class EncryptAesEcb {
-
-    public static Object encryptAesEcb(ArrayValue inputValue, ArrayValue keyValue,  Object padding) {
-        byte[] input = inputValue.getBytes();
-        byte[] key = keyValue.getBytes();
-        return CryptoUtils.aesEncryptDecrypt(CryptoUtils.CipherMode.ENCRYPT, Constants.ECB, padding.toString(), key,
-                                             input, null, -1);
-    }
-}
